Generate packs for each component
This lets us be more modular with what we ultimately send to the client.
The idea is to only send the client what it needs for a particular page.
It may be worth giving the browser hints that it should preload all the
packs so that they're already cached by the time the user wants them.